// WADNR.Web/src/app/shared/generated/model/project-contact-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectContactItemRequest } from './project-contact-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectContactSaveRequest { 
    Contacts?: Array<ProjectContactItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectContactSaveRequestForm { 
    Contacts?: FormControl<Array<ProjectContactItemRequest>>;
}

export class ProjectContactSaveRequestFormControls { 
    public static Contacts = (value: FormControlState<Array<ProjectContactItemRequest>> | Array<ProjectContactItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectContactItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}