// WADNR.Web/src/app/shared/generated/model/project-contact-item-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectContactItemRequest { 
    ProjectPersonID?: number | null;
    PersonID: number;
    ProjectPersonRelationshipTypeID: number;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectContactItemRequestForm { 
    ProjectPersonID?: FormControl<number>;
    PersonID: FormControl<number>;
    ProjectPersonRelationshipTypeID: FormControl<number>;
}

export class ProjectContactItemRequestFormControls { 
    public static ProjectPersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static PersonID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static ProjectPersonRelationshipTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
}